update:出力がない場合の対処を追加
